Split FirstViewController into separate files
